requirements.txt
# AWS Bedrock AgentCore and Strands
bedrock-agentcore-starter-toolkit>=0.1.0
strands-agents[otel]>=0.1.0
strands-agents-tools>=0.1.0
aws-opentelemetry-distro>=0.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Testing frameworks
pytest>=8.0.0
pytest-asyncio>=0.23.0
hypothesis>=6.98.0

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Data handling
pydantic>=2.5.0
dataclasses-json>=0.6.3

# HTTP and web
requests>=2.31.0
beautifulsoup4>=4.12.0

# Logging and monitoring
structlog>=24.1.0

# Date and time
python-dateutil>=2.8.2
